Don't override pending payments.

+++ libs/sdk-core/src/persist/transactions.rs
@@ -16,9 +16,15 @@ impl SqliteStorage {
     /// Note that, if a payment has details of type [LnPaymentDetails] which contain a [SuccessActionProcessed],
     /// then the [LnPaymentDetails] will NOT be persisted. In that case, the [SuccessActionProcessed]
     /// can be inserted separately via [SqliteStorage::insert_payment_external_info].
-    pub fn insert_or_update_payments(&self, transactions: &[Payment]) -> PersistResult<()> {
-        let deleted = self.delete_pending_lightning_payments()?;
-        debug!("Deleted {deleted} pending payments");
+    pub fn insert_or_update_payments(
+        &self,
+        transactions: &[Payment],
+        delete_pending: bool,
+    ) -> PersistResult<()> {
+        if delete_pending {
+            let deleted = self.delete_pending_lightning_payments()?;
+            debug!("Deleted {deleted} pending payments");
+        }
 
         let con = self.get_connection()?;
         let mut prep_statement = con.prepare(
@@ -526,8 +532,8 @@ fn test_ln_transactions() -> PersistResult<(), Box<dyn std::error::Error>> {
     }];
     let storage = SqliteStorage::new(test_utils::create_test_sql_dir());
     storage.init()?;
-    storage.insert_or_update_payments(&txs)?;
-    storage.insert_or_update_payments(&failed_txs)?;
+    storage.insert_or_update_payments(&txs, false)?;
+    storage.insert_or_update_payments(&failed_txs, false)?;
     storage.insert_payment_external_info(
         payment_hash_with_lnurl_success_action,
         Some(&sa),
@@ -587,7 +593,7 @@ fn test_ln_transactions() -> PersistResult<(), Box<dyn std::error::Error>> {
     let max_ts = storage.last_payment_timestamp()?;
     assert_eq!(max_ts, 2000);
 
-    storage.insert_or_update_payments(&txs)?;
+    storage.insert_or_update_payments(&txs, false)?;
     let retrieve_txs = storage.list_payments(ListPaymentsRequest::default())?;
     assert_eq!(retrieve_txs.len(), 3);
     assert_eq!(retrieve_txs, txs);
